requirements.txt
pillow
flask
spotipy
requests